Finalizando implementacao da rota de slots ocupados

+++ src/main/java/com/dev/trabProjarq/dominio/services/IOcupacaoAeroviaRep.java
@@ -6,5 +6,5 @@ import java.util.List;
 import com.dev.trabProjarq.dominio.entities.OcupacaoAerovia;
 
 public interface IOcupacaoAeroviaRep {
-    List<OcupacaoAerovia> findOcupadasSlots(int aeroviaId, Date data, List<Integer> slotsHorarios);
+    List<OcupacaoAerovia> findOcupadasSlots(int aeroviaId, Date data, List<Float> slotsHorarios);
 }
